+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [PR #309](https://github.com/rqlite/rqlite/pull/309): Tweak start-up logo.
 - [PR #308](https://github.com/rqlite/rqlite/pull/308): Move to clearer command-line options.
 - [PR #307](https://github.com/rqlite/rqlite/pull/307): Support node-to-node encryption. Fixes [issue #93](https://github.com/rqlite/rqlite/issues/93).
+- [PR #310](https://github.com/rqlite/rqlite/pull/310): HTTP service supports registration of Status providers; Mux is first client.
 
 ## 3.14.0 (May 4th 2017)
 - [PR #304](https://github.com/rqlite/rqlite/pull/304): Switch to Go 1.8.1.